requirements.txt
flask
sqlalchemy
openai
python-dotenv
loguru
fuzzywuzzy
python-Levenshtein
python-docx
pony
pyTelegramBotAPI
numpy
